Move notify.py to _internal
