+++ certbot/tests/notify_test.py
@@ -1,4 +1,4 @@
-"""Tests for certbot.notify."""
+"""Tests for certbot._internal.notify."""
 import socket
 import unittest
 
@@ -8,9 +8,9 @@ import mock
 class NotifyTests(unittest.TestCase):
     """Tests for the notifier."""
 
-    @mock.patch("certbot.notify.smtplib.LMTP")
+    @mock.patch("certbot._internal.notify.smtplib.LMTP")
     def test_smtp_success(self, mock_lmtp):
-        from certbot.notify import notify
+        from certbot._internal.notify import notify
         lmtp_obj = mock.MagicMock()
         mock_lmtp.return_value = lmtp_obj
         self.assertTrue(notify("Goose", "auntrhody@example.com",
@@ -18,10 +18,10 @@ class NotifyTests(unittest.TestCase):
         self.assertEqual(lmtp_obj.connect.call_count, 1)
         self.assertEqual(lmtp_obj.sendmail.call_count, 1)
 
-    @mock.patch("certbot.notify.smtplib.LMTP")
-    @mock.patch("certbot.notify.subprocess.Popen")
+    @mock.patch("certbot._internal.notify.smtplib.LMTP")
+    @mock.patch("certbot._internal.notify.subprocess.Popen")
     def test_smtp_failure(self, mock_popen, mock_lmtp):
-        from certbot.notify import notify
+        from certbot._internal.notify import notify
         lmtp_obj = mock.MagicMock()
         mock_lmtp.return_value = lmtp_obj
         lmtp_obj.sendmail.side_effect = socket.error(17)
@@ -32,10 +32,10 @@ class NotifyTests(unittest.TestCase):
         self.assertEqual(lmtp_obj.sendmail.call_count, 1)
         self.assertEqual(proc.communicate.call_count, 1)
 
-    @mock.patch("certbot.notify.smtplib.LMTP")
-    @mock.patch("certbot.notify.subprocess.Popen")
+    @mock.patch("certbot._internal.notify.smtplib.LMTP")
+    @mock.patch("certbot._internal.notify.subprocess.Popen")
     def test_everything_fails(self, mock_popen, mock_lmtp):
-        from certbot.notify import notify
+        from certbot._internal.notify import notify
         lmtp_obj = mock.MagicMock()
         mock_lmtp.return_value = lmtp_obj
         lmtp_obj.sendmail.side_effect = socket.error(17)
