--- docs/api/notify.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.notify`
--------------------------
-
-.. automodule:: certbot.notify
-   :members:
